# frontend/src/shared/api/api.yaml
openapi: 3.0.0
paths:
  /user/login/anonymous:
    post:
      operationId: UserController_loginAnonymous
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnonymousLoginDto'
      responses:
        '200':
          description: The anonymous user has been successfully logged in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
      summary: Anonymous user login
      tags:
        - User
  /user/login:
    post:
      operationId: UserController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: The user has been successfully logged in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
      summary: Registered user login
      tags:
        - User
  /user/register:
    post:
      operationId: UserController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: The user has been successfully registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
      summary: User registration
      tags:
        - User
  /user/me:
    get:
      operationId: UserController_me
      parameters: []
      responses:
        '200':
          description: The current user info has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeDtoResponse'
      security:
        - bearer: []
      summary: Get current user info
      tags:
        - User
  /user/{id}:
    get:
      operationId: UserController_getUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The user info has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserShortInfoResponseDto'
      summary: Get user by ID
      tags:
        - User
  /user/{id}/profile:
    get:
      operationId: UserController_getUserProfileById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The user profile has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
      summary: Get user profile by ID
      tags:
        - User
  /user/short-info:
    get:
      operationId: UserController_getUsersShortInfo
      parameters:
        - name: userIds
          required: true
          in: query
          description: Array of user IDs
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: The users short info has been successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserShortInfoResponseDto'
      summary: Get users short info
      tags:
        - User
  /user/avatar:
    put:
      operationId: UserController_updateAvatar
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadAvatarDto'
      responses:
        '200':
          description: Avatar has been successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadAvatarResponseDto'
      security:
        - bearer: []
      summary: Upload and update user avatar
      tags:
        - User
...
info:
  title: API Docs
  description: API Docs description
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    AnonymousLoginDto:
      type: object
      properties:
        name:
          type: string
          example: Guest_123
      required:
        - name
    UserDto:
      type: object
      properties:
        id:
          type: string
          description: Unique UUID
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        name:
          type: string
          description: User name
          example: Alex
        role:
          type: string
          enum:
            - anonymous
            - registered
          description: User role
      required:
        - id
        - name
        - role
    LoginResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          description: User info
          allOf:
            - $ref: '#/components/schemas/UserDto'
      required:
        - accessToken
        - user
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: password123
      required:
        - email
        - password
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        name:
          type: string
          example: John Doe
        username:
          type: string
          example: john_doe
        password:
          type: string
          example: password123
      required:
        - email
        - name
        - username
        - password
    MeDtoResponse:
      type: object
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          example: John Doe
        role:
          type: string
          example: registered
          enum:
            - registered
            - anonymous
        username:
          type: string
          example: john_doe
        email:
          type: string
          example: user@example.com
        avatarUrl:
          type: string
          example: https://example.com/avatar.jpg
      required:
        - id
        - name
        - role
    UserShortInfoResponseDto:
      type: object
      properties:
        id:
          type: string
          description: User's ID
        name:
          type: string
          description: User's full name
        username:
          type: string
          description: User's internal username
        avatarUrl:
          type: string
          description: User's avatar URL
      required:
        - id
        - name
        - username
        - avatarUrl
    UserProfileDto:
      type: object
      properties:
        id:
          type: string
          description: User ID
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Display name
          example: John Doe
        username:
          type: string
          description: Username
          example: johndoe123
        avatarUrl:
          type: string
          description: Avatar URL
          example: https://example.com/avatar.jpg
        totalGamesPlayed:
          type: number
          description: Total games played
          example: 42
        totalWins:
          type: number
          description: Total wins
          example: 15
        totalScore:
          type: number
          description: Total score
          example: 1250
      required:
        - id
        - name
        - username
        - avatarUrl
        - totalGamesPlayed
        - totalWins
        - totalScore
    UploadAvatarDto:
      type: object
      properties:
        avatar:
          type: string
          format: binary
          description: Avatar
      required:
        - avatar
    UploadAvatarResponseDto:
      type: object
      properties:
        newAvatar:
          type: string
          description: New avatar URL
      required:
        - newAvatar
...